Gijs Check of Template
Hey Lachlan, I think it looks great. I went through the assignment, run your solution codes, checked for typos (almost none). 
Questions are clear and straight to the point. Question 4 mentions 'this application', maybe mention shortly what application we're talking about (but this is kind of clear already as we're clearly discussing the paper, so do what you want with it)